requirements.txt
flask
flask-cors
pymongo[srv]
pyjwt
werkzeug
python-dotenv
google-generativeai
gunicorn
